adds attack action

+++ android/src/com/github/demongo/Demon.java
@@ -138,7 +138,7 @@ public class Demon implements Parcelable {
         return hp;
     }
 
-    public void setHp(int hp) {
+    public void setHp(long hp) {
         this.hp = hp;
     }
